requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-environ==0.11.2
pillow==10.3.0
redis==5.0.4
drf-yasg==1.21.7
pandas==2.2.2
openpyxl==3.1.5
celery==5.4.0
requests==2.32.3